Fix slow nested calls to 'embedded'.
It is common to nest calls to 'Proto3.Wire.Encode.embedded'.

But that function computed the length of the embedded message by
executing its builder to produce a temporary lazy ByteString.

Therefore when you run the builder returned by 'embedded', you
indirectly run its argument builder twice.  And if that builder
is itself produced by 'embedded', then the situation worsens.
The more deeply nested the data, the worse the performance.

This commit memoizes the number of bytes produced by each builder,
so that data structures being encoded are traversed only twice,
no matter the depth of nested calls to 'embedded': once to compute
the length of the encoding, and again to emit the encoding.

That is the theory, anyway.  Empirically, consider the test that
would be produced by applying the patch included below.  We observed
this commit to reduce the time required for that test from 3.42 seconds
to 0.01 seconds.  Of course, the nesting level is deeper than you would
see in typical practice, but it does illustrate the point.  Please note
that we are not actually checking in this test, because it never fails,
and adding some kind of performance testing framework seems like overkill.

diff --git a/proto3-wire.cabal b/proto3-wire.cabal
index bf33a16..3d4782a 100644
--- a/proto3-wire.cabal
+++ b/proto3-wire.cabal
@@ -44,3 +44,4 @@ test-suite tests
                        tasty-hunit >= 0.9 && <0.10,
                        tasty-quickcheck >= 0.8.4 && <0.9,
                        text >= 0.2 && <1.3
+  ghc-options:         -O2
diff --git a/test/Main.hs b/test/Main.hs
index cae12da..1ee130a 100644
--- a/test/Main.hs
+++ b/test/Main.hs
@@ -39,7 +39,9 @@ main = defaultMain tests

 tests :: TestTree
 tests = testGroup "Tests" [ roundTripTests
-                            , decodeNonsense ]
+                          , decodeNonsense
+                          , performance
+                          ]

 roundTripTests :: TestTree
 roundTripTests = testGroup "Roundtrip tests"
@@ -122,3 +124,10 @@ decodeNonsense :: TestTree
 decodeNonsense = HU.testCase "Decoding a nonsensical string fails." $ do
   let decoded = Decode.parse (one Decode.fixed64 0 `at` fieldNumber 1) "test"
   HU.assertBool "decode fails" $ isLeft decoded
+
+performance :: TestTree
+performance = HU.testCase "Performance." $ do
+  let builder 0 = Encode.byteString 0 "foo"
+      builder n = Encode.embedded (FieldNumber n) (builder (n - 1))
+  BL.foldrChunks seq (return ()) (Encode.toLazyByteString (builder 10000))
+  HU.assertBool "always succeeds" $ True