Fix `getFields` to handle interleaved repeated fields
`getFields` groups fields by key but does not sort by the key first,
meaning that if the keys are ordered like this:

```
[(k1, v1), (k2, v2), (k1, v3)]
```

... then the second occurrence of `k1` will be dropped instead of
collected with the first occurrence of `k1`.

In addition to fixing the algorithm to occur non-contiguous repetitions
this also speeds up the grouping code.  After benchmarking a wide
variety of algorithms to sort and group values by key, the fastest algorithm
was to use `Data.HashMap.fromListWith`.  This converts the `HashMap` to a
`Map` to preserve the original API and there is no loss of performance
due to this conversion.