Add support for running multiple parsers for support oneof keyword
it repeatdly try to find a matching field number, until one is found matching,
then this parser is called on the content.

* If no field matching is found then the first parser is used with mempty; this
  similar to what the `at` combinator is doing.
* If multiple matching fields exist, then the first matching one will take
  precedence, and the last field value matching this number will be used.
  This is (probably) different than what the protobuf spec specify, which
  seems to indicate that the last value of the group of possible field number
  should be used.

The right behavior lead to a much bigger change and non-self contain change
in the Decode module, and in the spirit of getting most of the feature
working, not done outright. the behavior difference should only be met
when explicitely adding different alternatives to the encoding message which
should be rare.