Add foldFields and don't pack the decodeWire into a Map
By keeping decodeWire in a list of parsed value, this keep the
ability to parse the stream of (FieldNumber, ParsedField) in the
order they appears in the field.

Add a combinator foldFields that fold over the stream of (FieldNumber,
ParsedField) and repeatdly try to call the associated parser then apply to
result into the accumulator. This allow the driven code to be
in control of the bias to use on a per field basis.