For speed, encode in reverse order. (#56)
Because protobuf wire format prefixes nested messages with
their lengths, we wish a fast way to compute those lengths.

A previous optimization was to memoize the encoded length of
a builder in order to avoid asymptotically poor performance,
but the cost of such memoization was materialization of trees
of builders corresponding to the data structures being encoded.

We now replace that memoization approach with a builder that
encodes in reverse order, tracking total bytes as it goes.
This approach allows us to save a snapshot of that total, encode
a suffix, resample that total, and take the difference to find
the length of the suffix.  We can then encode that length as
part of a prefix, because we are writing backward from the end.

Though memory is required to save the old total, that memory
is used just once, in stack fashion, and does not force all
of the builder to remain reachable on the heap.  And there
are other, miscellaneous improvements in how we use builders.

The effectiveness of this optimization will of course vary
with the data being serialized.  In one production setting
we observed an improvement of between 10% and 15%
in the speed of the serialization itself (with a much smaller
overall speed gain for the application as a whole).

This change also raises the minimum GHC version from 8.0 to 8.6
and discards support for Nix versions strictly less than 1.12.