1.4.4: Add reverse encoders for packed repeated fields.
Like the Vector encoders but unlike the forward-order
Foldable encoders they avoid the need to save elements
to the stack before encoding them, instead immediately
encoding the final element as the first step.  However,
the new encoders lack the special optimizations around
size checks that are possible for the Vector encoders,
which have advance knowledge of the number of elements.

This change also avoids compiler warnings in benchmark code.